Mr Whippy, a replication of an icecream truck
Turn your robot into a functioning Mr Whippy van. Find the
music for Fur Elise (which is the most common theme song
used) and get your robot to play this tune. Obviously if you
have a customer, you need to stop (use supersonic sensor).
Every time your robot stops, the music pauses, and a light
flashes as a safety measure. Restart Mr Whippy again by
pressing the touch sensor, the music restarts and the light
stops flashing. When you come to a crossing, use a light
sensor to be aware of this and get your robot to stop for 5
seconds. The rest I’ll leave up to you. Make your robot
behave in as realistic a way as possible – ie* interaction with
customers, graphics/text that appears on the screen etc....