Making the robot dance to the Nutbush
Create a dance routine for your robot to the music of the classic song – “Nutbush City Limits” by Tina Turner. Your robot should remain within a 1 metre square area. You are not required to use sensors for this task. Like most songs, “Nutbush City Limits” follows a pattern of verses and a chorus. Demonstrate your programming skills to make your
code as efficient as possible – use repeat style commands and separate sub task style commands to achieve this goal. Include turns, and forwards and backwards movements using the appropriate commands.

Incorporate flashing lights – or use an additional motor for some other purpose. Your robot should dance for a minute of the song. You should try to synchronise the dance movements with the song. You will need to develop a sequence of coded robot movements that gets your robot to move in time and with energy. Incorporate lights that flash at significant times in the song. Use subroutines to structure the movements into different distinct sections. Your main task section should be a sequence of calls to these subroutines. Use this question as an opportunity to demonstrate your ability to effectively use constants, variables and the repeat command. Your routine should not be so simple that it doesn’t allow you to demonstrate your coding ability.